Added redirect host filtering.
With the rule-based 'allow-redirect' boolean is set to true, this allows for a
downstream to reply with a status code of 377 and a header 'x-internal-redirect'
to have a value of a hostname which the proxy will use instead of the current
associated downstream.

This could be a potential problem if the downstreams configured are not trusted
by the administrator, and a downstream could redirect requests to a potentially
malicious server.

A new configuration setting has been added to aid in this type of activity:

rule x {
	allow-redirect  = true
	redirect-filter = { "*host1.com, "www.yahoo.com" }
}

In the above rule, only redirects matching the wildcard *host1.com or the
specific value of "www.yahoo.com" will be accepted. If the redirect value is not
a host that is contained within this list, a 401 response will be sent to the
upstream client.

If the allow-redirect is true and there is no redirect-filter configuration set,
RProxy will allow any redirect responses.