Close connection if no rules are matched.
This fixes an issue where a client can make a request for a rule that does not
exist in a vhost, and by default none of the rproxy callbacks would be called.
So if a malicious user were to post a large amount of data, this could cause a
DoS.

This sets a default callback for rules which is a catch-all that immediately
dumps the connection.